// frontend-svc/frontend/src/components/Tabs/Tabs.module.scss.d.ts
// Generated file, DO NOT EDIT. Check DEV.md on how to regenerate
export type Styles = {
  multi_tabs: string;
  selected: string;
  tab__item: string;
  tab__item__link: string;
  tab__item__link__content: string;
  tab__item__link__prefix: string;
  tab__item__link__suffix: string;
  with_suffix: string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;